Fase 1 Python SQL/requirements.txt
python-dotenv
psycopg2-binary
pytest